feat(types): Global typings
