+++ module/rest-express/src/global.d.ts
@@ -3,17 +3,13 @@ import * as rest from '@travetto/rest';
 import * as express from 'express';
 
 // Support typings
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace Express {
-    // eslint-disable-next-line no-shadow
-    interface Request {
-      [TravettoEntitySymbol]?: rest.Request;
-      [NodeEntitySymbol]?: express.Request;
-    }
-    interface Response {
-      [TravettoEntitySymbol]?: rest.Response;
-      [NodeEntitySymbol]?: express.Response;
-    }
+declare module 'express' {
+  interface Request {
+    [TravettoEntitySymbol]?: rest.Request;
+    [NodeEntitySymbol]?: express.Request;
+  }
+  interface Response {
+    [TravettoEntitySymbol]?: rest.Response;
+    [NodeEntitySymbol]?: express.Response;
   }
 }
