+++ module/runtime/src/global.d.ts
@@ -6,4 +6,29 @@ declare global {
   interface WritableStreamDefaultWriter<W = any> {
     [write]?: (a: W) => void;
   }
+
+  /**
+   * @concrete node:buffer#Blob
+   */
+  interface Blob { }
+
+  /**
+   * @concrete node:buffer#File
+   */
+  interface File { }
+
+  namespace NodeJS {
+    /**
+     * @concrete node:stream#Readable
+     */
+    interface ReadableStream { }
+  }
 }
+
+declare module 'stream' {
+  /**
+   * @concrete node:stream#Readable
+   */
+  interface Readable { }
+}
+
