+++ module/schema/src/global.d.ts
@@ -9,28 +9,4 @@ declare global {
      */
     from<T>(this: Class<T>, data: DeepPartial<T>, view?: string): T;
   }
-
-  /**
-   * @concrete node:buffer#Blob
-   */
-  interface Blob { }
-
-  /**
-   * @concrete node:buffer#File
-   */
-  interface File { }
-
-  namespace NodeJS {
-    /**
-     * @concrete node:stream#Readable
-     */
-    interface ReadableStream { }
-  }
-}
-
-declare module 'stream' {
-  /**
-   * @concrete node:stream#Readable
-   */
-  interface Readable { }
 }
